timer, nats: allow for disk storage so nodes can recover from a crash
Use levelup wrapper to allow for a swappable persistence layer. Defaults
to memdown for in memory storage. Timer class will now write to leveldb
first before setting a timer and clear it when the callback is
triggered.

If a node crashes before it can rebalance, the next time it starts, it
will read all values from the db, and set the timer as if it were an
incoming request. Stale timers are executed immediately, etc.

replace mocha w/ tap it is easier to test things with tap.
Code coverage is baked in. And I can just execute a file,
which is impossible with mocha

constructor accepts a callback to execute once the recovery process
has completed.

Semver: major

major