skyring, server: sets default behavior to not auto rebalance
The previous behavior would send every server in the cluster into a
rebalance every time a server was added or removed from the cluster.
This can lead to some undesireable bahavior.

Adds a new option `autorebalance` that can be enabled if that behavior
is desired. By default a server will rebalance when it leaves the
cluster.

Additionally, the linux signal SIGUSR2 can be sent to the parent process
to trigger a specific node to rebalance

Semver: major